requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
flask>=2.0.0
pyyaml>=5.4.0
tqdm>=4.60.0
plotly>=5.0.0
rdkit-pypi>=2021.9.0
pymatgen>=2022.0.0
ase>=3.22.0